+++ packages/react-instantsearch/examples/react-native/index.android.js
@@ -5,5 +5,5 @@
  */
 
 import { AppRegistry } from 'react-native';
-import App from './App';
+import App from './src/App';
 AppRegistry.registerComponent('reactnative', () => App);
