+++ packages/react-instantsearch/examples/react-native/index.ios.js
@@ -5,6 +5,6 @@
  */
 
 import { AppRegistry } from 'react-native';
-import App from './App';
+import App from './src/App';
 
 AppRegistry.registerComponent('reactnative', () => App);
