+++ packages/react-instantsearch/examples/react-native/src/App.js
