+++ packages/react-instantsearch/examples/react-native/src/components/Highlight.js
